requirements.txt
mysql-connector-python==9.6.0
Flask==2.3.3
flask-cors==4.0.0